// src/types/User.ts
export interface User {
    id: number;
    username: string;
    password: string;
    status: boolean;
    name: string;
    phone: string;
    address: string;
    email: string;
    cccd: string;
    gender: number;
    birthDate: string;
    role: number;
    licenseCategory: string;
    expirationDate: string;
}